Add sink-specific computed values to log stream
The `awsPartnerEventSource` and `azurePartnerTopic` properties are
returned as part of the `sink` property returned by the Auth0 API if the
`type` is 'eventbridge' or 'eventgrid', respectively.

As these values can be useful for output to the user, or for
provisioning other resources in Terraform, it's useful to make these
elements available as part of the resource schema.